Use buildid as a file key for uprobes instead of inode + device
(device, inode) is not a correct key for uprobe resolving. device might be different for the same file if it is accessed via different mount namespaces. This happens if we specify binaries path for uprobes in format `/proc/<pid>/root/<container_path>` - the device might be different from the device that we will find in agent mapping scan because of overlay fs.

Simple solution - use buildid as a file key instead of inodes and devices.
commit_hash:35d5cc7cec64beccfd072ffcc915bcd03996f59c